small opts

+++ include/defs.hpp
@@ -4,7 +4,7 @@
 #include <tuple>
 #include <vector>
 
-using ScoreType = int;
+using ScoreType = int16_t;
 
 using ScoreMatrix = Eigen::Matrix<ScoreType, -1, -1, Eigen::ColMajor>;
 using TraceMatrix = Eigen::Matrix<char, -1, -1, Eigen::ColMajor>;
